fix long text oom using ssml; filter comma; update polyphonic
