+++ paddlespeech/t2s/frontend/polyphonic.yaml
@@ -49,3 +49,6 @@ polyphonic:
     扎实: ['zha1','shi2']
     干将: ['gan4','jiang4']
     陈威行: ['chen2', 'wei1', 'hang2']
+    郭晟: ['guo1', 'sheng4']
+    中标: ['zhong4', 'biao1']
+    抗住: ['kang2', 'zhu4']
