+++ paddlespeech/t2s/frontend/ssml/xml_processor.py
@@ -1,4 +1,17 @@
 # -*- coding: utf-8 -*-
+# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import re
 import xml.dom.minidom
 import xml.parsers.expat
@@ -68,7 +81,8 @@ class MixTextProcessor():
             after_xml = mat.group(3)
 
             # pre with none syllable
-            ctlist.append([pre_xml, []])
+            if pre_xml:
+                ctlist.append([pre_xml, []])
 
             # between with syllable
             # [(sub sentence, [syllables]), ...]
@@ -77,9 +91,11 @@ class MixTextProcessor():
             ctlist = ctlist + pinyinlist
 
             # post with none syllable
-            ctlist.append([after_xml, []])
+            if after_xml:
+                ctlist.append([after_xml, []])
         else:
             ctlist.append([mixstr, []])
+
         return ctlist
 
     @classmethod
@@ -94,15 +110,18 @@ class MixTextProcessor():
             in_xml = mat.group(2)
             after_xml = mat.group(3)
 
-            ctlist.append(pre_xml)
+            if pre_xml:
+                ctlist.append(pre_xml)
+
             dom = DomXml(in_xml)
             tags = dom.get_text_and_sayas_tags()
             ctlist.extend(tags)
 
-            ctlist.append(after_xml)
-            return ctlist
+            if after_xml:
+                ctlist.append(after_xml)
         else:
             ctlist.append(mixstr)
+
         return ctlist
